moved music control to different file
